feat: connect custom logger to root app

+++ src/logger/logger.service.ts
@@ -1,6 +1,6 @@
 import 'dotenv/config';
 import { ConsoleLogger, LoggerService, Injectable } from '@nestjs/common';
-import { appendFile, stat, rename, writeFile } from 'fs/promises';
+import { appendFile, stat, rename, writeFile, access } from 'fs/promises';
 import * as path from 'node:path';
 import { ConfigService } from '@nestjs/config';
 
@@ -28,17 +28,24 @@ export class MyCustomLoggingService implements LoggerService {
     const maxSize = +this.configService.get<string>('MAX_LOGS_SIZE');
 
     if (size >= maxSize) {
-      await rename(
-        this.logPath,
-        path.join(
-          __dirname,
-          '..',
-          '..',
-          'custom-logs',
-          `logs.${Date.now()}.txt`,
-        ),
-      );
-      await writeFile(this.logPath, '');
+      const isExist = async (path: string) =>
+        await access(path)
+          .then(() => true)
+          .catch(() => false);
+
+      if (await isExist(this.logPath)) {
+        await rename(
+          this.logPath,
+          path.join(
+            __dirname,
+            '..',
+            '..',
+            'custom-logs',
+            `logs.${Date.now()}.txt`,
+          ),
+        );
+        await writeFile(this.logPath, '');
+      }
     }
 
     await appendFile(this.logPath, message);
